Memory Pod, Newton Room, Fader, and Level Loader
Clarks additions are working. The shader based fade system is not. There is a new Teleport Area With Update script for ground floors. The delay is working, but not the actually fading. Also brought in our old Level Loader script from Hull Breach and the Crossfade canvas to fade the old (bad) way.