Socket alignment
Got an alignment check built in to the Interface state switcher. Now, when the tip of the needle collides with the inside end of the Interface Socket, a check is done to see if the angle is aligned close enough to an empty "Socket End" object. This will now prevent Injector occurring if the needle isn't properly aligned inside the Socket.