New docking method
Changed how the Injector works a bit. Now, upon docking, the current Injector goes away, and the Persistent Injector (from the Persistent scene) re-enables. This is done so that upon docking, the Injector stays "locked" in and moves with the Left Hand. Note: the needle tip collider material is still seen during transisitons.